requirements.txt
PyOpenGL
numpy
pygame
Pillow